Move experiences to sass modules
