+++ app/frontend/Pages/index.ts
@@ -1,7 +1,7 @@
 export { default as Home } from './home';
 export { default as About } from './about';
 export { default as Join } from './join';
-export { default as Experience } from './Experience';
+export { default as Experience } from './Experience/Experience';
 export { default as Stylesheet } from './stylesheet';
 export { default as Create } from './Create/Create';
 export { default as Register } from './Register';
